Change tests to work after merging main, improve Python filter_map
