+++ iam-policy-autopilot-policy-generation/src/extraction/python/extractor.rs
@@ -100,22 +100,22 @@ impl Extractor for PythonExtractor {
         service_index: &ServiceModelIndex,
     ) {
         let method_disambiguator = MethodDisambiguator::new(service_index);
+        let resource_extractor = ResourceDirectCallsExtractor::new(service_index);
+        let waiters_extractor = WaitersExtractor::new(service_index);
+        let paginator_extractor = PaginatorExtractor::new(service_index);
 
         for extractor_result in extractor_results.iter_mut() {
             match extractor_result {
                 ExtractorResult::Python(ast, method_calls) => {
                     // Extract resource direct calls (with ServiceModelIndex access)
-                    let resource_extractor = ResourceDirectCallsExtractor::new(service_index);
                     let resource_calls = resource_extractor.extract_resource_method_calls(ast);
                     method_calls.extend(resource_calls);
 
                     // Add waiters to extracted methods using the service model index directly
-                    let waiters_extractor = WaitersExtractor::new(service_index);
                     let waiter_calls = waiters_extractor.extract_waiter_method_calls(ast);
                     method_calls.extend(waiter_calls);
 
                     // Add paginators to extracted methods using the service model index directly
-                    let paginator_extractor = PaginatorExtractor::new(service_index);
                     let paginator_calls = paginator_extractor.extract_paginate_method_calls(ast);
                     method_calls.extend(paginator_calls);
